rename data-integrity-v1 to data-integrity
