+++ packages/askar-to-drizzle-storage-migration/tests/migrate.drizzle.e2e.test.ts
@@ -26,7 +26,7 @@ import { askar, askarPostgresStorageConfig } from '../../askar/tests/helpers'
 import didKeyP256 from '../../core/src/modules/dids/__tests__/__fixtures__/didKeyP256.json'
 import { sprindFunkeTestVectorBase64Url } from '../../core/src/modules/mdoc/__tests__/mdoc.fixtures'
 import { sdJwtVcWithSingleDisclosure } from '../../core/src/modules/sd-jwt-vc/__tests__/sdjwtvc.fixtures'
-import { CredoDidKeyDiVcEncoded } from '../../core/src/modules/vc/data-integrity-v1/__tests__/fixtures/credo-di-vc'
+import { CredoDidKeyDiVcEncoded } from '../../core/src/modules/vc/data-integrity/__tests__/fixtures/credo-di-vc'
 import { CredoEs256DidJwkJwtVc } from '../../core/src/modules/vc/jwt-vc/__tests__/fixtures/credo-jwt-vc-v2'
 import { customDocumentLoader } from '../../core/src/modules/vc/linked-data-proofs/__tests__/documentLoader'
 import { Ed25519Signature2018Fixtures } from '../../core/src/modules/vc/linked-data-proofs/__tests__/fixtures'
