+++ packages/core/src/modules/vc/data-integrity/W3cV2DataIntegrityIssueMapping.ts
