+++ packages/core/src/modules/vc/data-integrity/__tests__/fixtures/credo-di-vc.ts
